requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
psycopg2-binary>=2.9
